Reorganize code according to code review
* tests moved to test for consistency with other timeflux repos
* graph moved to test/graph
* amti exception derives timeflux exception
* default values for driver constructor
* driver moved to timeflux_amti.nodes module
