+++ timeflux_amti/exceptions.py
@@ -1,6 +1,8 @@
 """Timeflux-AMTI exceptions"""
 
+from timeflux.core.exceptions import TimefluxException
 
-class TimefluxAmtiException(Exception):
+
+class TimefluxAmtiException(TimefluxException):
     """Exception thrown when there is a problem with the AMTI device """
     pass
